requirements.txt
click==8.0.3
itsdangerous==2.0.1
Pillow==8.4.0
PyQt5==5.15.5
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt5-stubs==5.15.2.0
Werkzeug==2.0.2
